refactor(serviceWorker): change file for folder core
