+++ src/core/serviceWorker.js
